btf: Add support for `bpf_core_type_matches()`
This commit adds support for the latest edition to CO-RE. The
`bpf_core_type_matches()` function allows you to check if a given type
matches. This is a stricter check than the normal compatibility check.

An example use case for this feature is to implement fallback code for
cases where kernel types changed over time, such as the
`block_rq_insert` tracepoint. The tracepoint lost an argument in the
v5.11 kernel, so this feature can be used to handle the change in
provided context in a CO-RE manner.

Signed-off-by: Dylan Reimerink <dylan.reimerink@isovalent.com>