ringbuf: simplify ringReader
The code to correctly access the ring is currently split across two
files an multiple functions which makes it hard to grasp what is going
on. Strip out the overly generic Reader interface and only expose
ringReader.readRecord. Also move discarding of samples into readRecord.

This removes the amount of code and cuts down on the number of places
where we call into sync/atomic.

Signed-off-by: Lorenz Bauer <lmb@isovalent.com>