btf: export As and remove Transformer
BTF encodes C language constructs such as const, restrict, etc. as
separate types. In most cases we don't care about these when traversing
a type graph. Users can strip these qualifiers by invoking Copy() with a
"transformer" aka UnderlyingType(). Unfortunately, copying is quite
expensive.

For this reason we changed CO-RE relocations to use an unexported as()
function that can be used to "unwrap" a type instead of copying it.
Think of it as a generic version of UnderlyingType. This is much faster
than copying and turns out to be quite ergonomic as well, since we often
have to assert a type anyways.

Export As() so that external users can benefit from it. Usage is like so:

    foo, ok := As[*Int](typ)
    if !ok {
        panic("not an Int")
    }

Remove the Transformer type and the extra argument from Copy, since As
is the better way to deal with qualifiers and typedefs. This allows
further simplifying Copy in a follow up commit.

Signed-off-by: Lorenz Bauer <lmb@isovalent.com>