use binary.NativeEndian
Remove the internal implementation of NativeEndian in favour of
binary.NativeEndian. There is one hiccup: it's not possible to
use a simple comparison to figure out whether NativeEndian is
the same as for example LittleEndian. The IsNativeEndian helper
takes care of this.

See https://github.com/golang/go/issues/57237#issuecomment-1361852721

Signed-off-by: Lorenz Bauer <lmb@isovalent.com>