tests: remove !IsNativeEndian skips
We have a lot of tests which first glob a list of files containing
BPF for little and big endian, then parse those files and finally
skip a test if the files endianness doesn't match the native
endianness.

This is pretty wasteful, since we parse BPF for no good reason and
makes the tests more verbose than they have to be. Use NativeFile
instead and remove almost all checks for IsNativeEndian from tests.

The two remaining cases are in the ELF loader tests, where they do
have value: we want to make sure that we can parse things from any
endianness.

Signed-off-by: Lorenz Bauer <lmb@isovalent.com>