elf_reader_test: Fix CI by disabling the "test_log_fixup" selftest
The "test_log_fixup" selftest is broken on purpose, with the goal of
testing error messages. Due to changes in #1364 we now emit a poison
instruction in `bad_relo_subprog` on a instruction which also has a
reference to a BPF-to-BPF function. This causes the symbol/reference
resolution step to fail.

This error is flaky, it only happens when `bad_relo_subprog` is loaded
first. If any other program in the collection is loaded first due to
random hash map ordering we skip the test instead due to the test
checking for verifier errors.

Since this test is broken on purpose, we previously just always hit this
skip condition. This commit disabled this selftest explicitly which
resolves the CI flakiness.

Signed-off-by: Dylan Reimerink <dylan.reimerink@isovalent.com>