program: remove fallthrough from error heuristics
We have a bunch of heuristics that try to give a better error when
loading a program fails. It's kind of tricky to decide which order
they should be in and what interactions they may have. Commit
148c76c refactored the check for missing bpf2bpf call support into a
switch with a fallthrough statement. As a result, we apply all of
our EINVAL heuristics to EPERM as well, even though the initial
intent was to only only share the bpf2bpf code.

Fix this by pulling the heuristic into it's own code block again/
The check is kind of expensive since it iterates all instructions
in the worst case so we also move it to the very end.

Signed-off-by: Lorenz Bauer <lmb@isovalent.com>