btf: add a workaround for the "No type found" error
The kernel accepts line info for each instruction in a program,
which is output as part of the verifier log and therefore provides
very important context. The strings for each line info are interned
into a BTF blob.

Unfortunately, the kernel refuses to load BTF with a string table but
without any types. This means it's not possible to add line info to
a program which doesn't use BTF otherwise.

Add a MarshalOption which fixes this behaviour by adding a dummy type
if necessary.

Signed-off-by: Lorenz Bauer <lmb@isovalent.com>