program: return better errors on CO-RE or kfunc fixup failures
CO-RE relocations and kfunc fixups can both fail for a number of
reasons. However, we can't return an error outright, since a program
may be written in a style that takes this into account.

Instead, we rewrite the BPF instruction with the failing relocation
or fixup to a call instruction to a magic number. This causes the
verifier to bail out if the program does end up trying to execute the
"invalid" instruction.

Add a heuristic which detects these cases from the verifier log and
returns a more intuitive error instead.

Signed-off-by: Lorenz Bauer <lmb@isovalent.com>