btf: synthesise instruction comments into line info
When an asm.Comment is added to an instruction, existing source
information like btf.Line will be replaced. But in some cases,
instructions without line information won't pass the verifier. The
expected behavior would be, that the comment gets synthesized into a
line info where everything but the line string is zero. This commit
implements this.

Fixes #1413

Signed-off-by: Marcus Wichelmann <mail@marcusw.de>