Extend go test CI from 10 to 15 miniutes
The default builtin timeout in the go test runner is 10 min and the GH runner is also limited to 10 min. In practice the GH runner is quicker to kill tests that exceed this time limit, in which case we abruptly exit without logs on which test caused the timeout.

This change extends the runner timeout to 15 min to give the go runner a chance to kill the bad test and to tell us which test was taking longer so that we can fix the issue.

Signed-off-by: Dylan Reimerink <dylan.reimerink@isovalent.com>