perf: do not overwrite named return
When calling newPerfEventRing() a new variable named err is created
which conflicts with the named return of the function
NewReaderWithOptions(). As a consequence the defer function to clean up,
that works on the named return of the function, can not work correctly
if an error is returned by newPerfEventRin() and may leak unclosed FDs.

Signed-off-by: Florian Lehner <dev@der-flo.net>