perf: clean up error handling and bare fds
The perf reader was written before we had the sys.FD abstraction
and therefore still uses plain int to keep track of fds. This means
we don't get leak tracking during CI. It also relies on a subtle
interaction between defer and named return values which is easy to
break by accidentally shadowing err:

    func ... (err error) {
        poller, err := poller.New(...)

        defer func() {
            if err != nil {} // Refers to the second err
        }()
    }

Rework the code to use sys.FD and the closeOnError idiom used
elsewhere in the library.

Signed-off-by: Lorenz Bauer <lmb@isovalent.com>