features: Fix test `TestHaveProgramType/Extension` in kernel >=v6.8
In kernel v6.8, logic was added that restricts valid func info BTF for
programs that can be replaced with an BPF_PROG_TYPE_EXT program.
Before this change the parameters of the extension program had to be
the same as the original program.

Commit https://github.com/torvalds/linux/commit/5eccd2db42d77e357 made
it so the verifier asserts that the BTF func info of "main" programs
must have exactly 1 parameter of a pointer type to the context struct,
even if the code doesn't use the context like in our case.

Signed-off-by: Dylan Reimerink <dylan.reimerink@isovalent.com>