Add get methods for name of program and map
When working with objects loaded from fds, it can be useful to have
access to their names. This is already possible via their *Info structs.
The objects themself have a name field, though, which are provisioned
with the *Info's value. So, the name is already present in the objects,
but not exposed yet.

This commit adds methods to Program and Map to access the value of their
name field.

Signed-off-by: Tobias Böhm <code@aibor.de>