link: add test suite for netkit
The netkit link support was added without a test suite due to a lack of support
for driver-specific extractions in the `github.com/jsimonetti/rtnetlink` package.
The driver support has been added to `rtnetlink`.  This commit adds a test suite
for the netkit link and pulls in the `github.com/jsimonetti/rtnetlink/v2` package for testing.

Signed-off-by: Birol Bilgin <birolbilgin@gmail.com>