feat(bpf2go): Export go binding generation
Moves the code to generate go bindings to its own dedicated package
and export the necessary functions to use it in bpf2go.

This change will make it possible to reuse this code in other tools,
without having to compile the C programs.

The behaviour of bpf2go is unchanged.

Two additional changes were made:
1. The 'Module' parameter has been made an input to the code generation.
  Otherwise embedding the bindings package in another module would lead to
  incorrect import statements in the generated code.
  I fail to see the value of the 'debug' lookup, the only valid value is
  "github.com/cilium/ebpf", but maybe there are some cilium internal build
  constraints, so I left it configurable.
2. The function which transforms type names to go identifiers has been updated
  to deal with all caps names. These shouldn't be present in C code, but
  appear when using rust. The change ensures the field names in the bindings
  are idiomatic Go field names.

Signed-off-by: Wouter Dullaert <wouter.dullaert@exoscale.ch>