bpf2go: export binding generator
Moves the code to generate go bindings to its own dedicated package
and export the necessary functions to use it in bpf2go.

This change will make it possible to reuse this code in other tools,
without having to compile the C programs.

The behaviour of bpf2go is unchanged.

[ Lorenz: move code and backed out changes around module and identifier.
  These can come later, with tests. ]

Co-authored-by: Lorenz Bauer <lmb@isovalent.com>
Signed-off-by: Wouter Dullaert <wouter.dullaert@exoscale.ch>