map: fix flaky TestMapIteratorAllocations
For some reason the allocation check is flaky, with an allocation
happening in sysenc.Unmarshal. Pinpointing the exact cause is difficult
because we can't tell whether the allocation happened during the warm
up or not.

Increase the number of runs for the allocation check, which is probably
a sensible thing to do anyways.

Signed-off-by: Lorenz Bauer <lmb@isovalent.com>