fix a variety of Copy() problems
The are multiple Copy() methods in the code base which are used to
create deep copies of a struct. Any bugs in them can manifest as
data races in concurrent code.

Add a quicktest checker which ensures that two variables are a
deep copy of each other: values match, but all locations in memory
differ.

Fix a variety of problems with the existing Copy() implementations.
They all allow copying a nil struct and copy all elements.

There are exceptions to the deep copy rule:

- MapSpec.Contents is not deep copied because we can't easily make
  copies of interface values. This is documented already.
- MapSpec.Extra is an immutable bytes.Reader and therefore only
  needs a shallow copy.
- ProgramSpec.AttachTarget is a Program, which is (currently) safe
  for concurrent use.

Fixes #1517

Signed-off-by: Lorenz Bauer <lmb@isovalent.com>