internal/sys: generate constants from unnamed enums
bpf.h defines a bunch of constants as unnamed enums. This makes them
tricky to pull out of vmlinux BTF, so instead we rely on the internal/unix
shim. This causes a small maintenance burden whenever we need a new flag,
but more importantly it makes some upcoming cross platform work harder.

Pull the constant definitions out of vmlinux BTF instead of relying on
package unix. Unfortunately a bunch of the flags are re-used.
This is most noticeable for MapFlags, for which the enum in the kernel
has become a bit of a grab bag. Instead of trying hard to wrap everything
into a neat type we just generate untyped constants and get rid of MapFlags.

Signed-off-by: Lorenz Bauer <lmb@isovalent.com>