CI: cross build for Windows instead of Darwin
CI cross compiles the package and its tests for Darwin to make sure
that it's possible to edit the source code on other Unix like platforms.

Up the game a little and cross build for Windows instead of Darwin:
this flushes out accidental uses of x/sys/unix instead of internal/unix.

Signed-off-by: Lorenz Bauer <lmb@isovalent.com>