New architecture: wasm
cilium/ebpf currently does not compile to wasm with tinygo. This patch
makes it compile.

I am working on a website where users can submit ebpf binaries (either
in ELF format or as an Inspektor Gadget export) and the website will
parse the ebpf binary client-side with wasm. The wasm code is written in
Go, using cilium/ebpf and is compiled with tinygo.

My wasm code uses ebpf.LoadCollectionSpecFromReader() to display
information about the ebpf binary. But it will not call
ebpf.NewCollection() because the wasm/javascript environment in the
browser cannot interact with the Linux kernel.

Signed-off-by: Alban Crequy <albancrequy@linux.microsoft.com>