add test to assert MapIterator failure against keyless (queues) maps
This commit introduces the `TestIterateWrongMap` test, which should
assert that the `MapIterator` fails while called on keyless maps such as
queues. We do not expect different behaviours, as the `MapIterator.Next`
relies on the presence of a key for which it lookups the value.
A different method can be later introduces to also traverse a queue, but
in that case we'd need to additionally remove the value.

Signed-off-by: Simone Magnani <simone.magnani@isovalent.com>