feat(gen): Make Identifier function configurable
The standard Identifier function used by the generator results in very
unidiomatic go identifier names when used with my ebpf object.
With this change, the Identifier function becomes configurable, allowing
me to pass in something that produces better outputs in my context.
If no function is passed in, we keep using the default, this makes the
change entirely backwards compatible.
Furthermore, a test has been added to ensure the configuration is
correctly passed through.

Signed-off-by: Wouter Dullaert <wouter.dullaert@exoscale.ch>