ebpf: interact with global variables using CollectionSpec.Variables
This commit adds a new field to CollectionSpec: Variables. Along with it,
a new type is introduced, the VariableSpec. It acts as a view into a data
section representing global variables and constants. They can be get and
set freely before loading the CollectionSpec into the kernel, after which
they will be converted into a new type, a Variable, which will be added
in a follow-up commit.

Like ProgramSpec and MapSpec, a VariableSpec can be assigned to a struct
using CollectionSpec.Assign. Should the programmer wish to hide certain
variables from being exposed in the CollectionSpec, they can be marked
using the 'hidden' C visibility attribute.

With this commit, all data sections are now included in the CollectionSpec
tests, while they were excluded before. This ensures a consistent experience
across LLVM versions.

Much of the code was borrowed from RewriteConstants, which will be turned
into a wrapper around CollectionSpec.Variables in a follow-up commit.

Signed-off-by: Timo Beckers <timo@isovalent.com>