link: tracefs: expect EINVAL when creating tracefs kprobes against missing symbols
As highlighted in the code, Linux commit 9d8616034f16 likely introduced a bug that
causes writes to <tracefs>/kprobe_events to fail with EINVAL if the target symbol
doesn't exist. Update the tests to reflect that behaviour. At least now it's on par
with kretprobes..

Also removed a tautological condition in NewEvent.

Signed-off-by: Timo Beckers <timo@isovalent.com>