ebpf: minimal windows support WIP
Add minimal support for Windows. Most APIs return ErrNotSupportedOnOS.
The only way to obtain Map and Program is by passing LoadCollection the
path to a signed Windows driver as is done in TestLoadNativeImage.

In theory it should be possible to JIT a program via NewProgram as it is
done in TestLoadXDPTestProgram but the test fails due to an RPC error.
This is most likely because ebpfsvc.exe is not running.

All API on Map and Program are untested / presumed not working, with the
exception of Info(). Info() works enough to create Map and Program, but
upstream efW has extended the definition of sys.MapInfo in an incompatible
way.

Signed-off-by: Lorenz Bauer <lmb@isovalent.com>