ebpf: distinguish between supported platforms WIP
Linux and Windows share the concept of program types, map types and so on.
For example, XDP is supported on both platforms. Unfortunately the constant
values used by both platforms are different. On Linux, XDP has value 0x6
while on Windows it is 0x1.

This problem extends to a CollectionSpec parsed from an ELF. Because of the
different values a Windows PerCPUHash aliases with a Linux ProgramArray.
This causes all sorts of mayhem that is hard to avoid without making too
many changes to the code base.

Introduce a Platform type which enumerates the supported platforms.
That type is then used to tag constant values. This means that there are
distinct constants even when a type exists on both platforms. Instead of
overloading the XDP constant we introduce WindowsXDP.

- Need to address asm.BuiltinFunc

Signed-off-by: Lorenz Bauer <lmb@isovalent.com>