internal/fdtrace: rename to testmain and add windows trace logging
The eBPF for Windows runtime logs traces and error messages to the
Windows event subsystem. Being able to correlate test failures with
this trace is very useful to diagnose problems.

On Windows, start an event tracing session before starting the tests,
and dump a summary to stderr after tests have finished.

The summary looks roughly like this:

      base       _get_map_descriptor_properties returned success
      entry-exit _ebpf_core_protocol_map_delete_element
    ! base       _ebpf_core_protocol_map_delete_element returned error   Error=19

The exclamation mark in the first column indicates an error.

Signed-off-by: Lorenz Bauer <lmb@isovalent.com>