internal: add a way to write portable feature tests
Feature tests are used to enable new features in a backwards compatible
manner. They exercise very specific code paths in the kernel. It makes
sense to assume that the existing feature tests will not work as intended
when porting to a new platform. Change the existing NewFeatureTest
constructor to return a new sentinel error when invoked on non-Linux OS.

Add a new constructor NewPortableFeatureTest which allows indicating on
which platforms and versions a feature test is reliable.

Signed-off-by: Lorenz Bauer <lmb@isovalent.com>