link: minimal windows support WIP
We retain the Link interface and the RawLink type, but all other link
types are not available on Windows. There is a possibility that we can
extend the XDP link to Windows, but that will have to come later.

Attaching a link has to call into the efW runtime directly since
the interface can't be mapped to Linux's BPF_LINK_CREATE:

- Program and attach types are represented as GUID, which don't fit
  the 32bit allocated by Linux.

Testing the Link attach code is cumbersome since the only way to load
a program via a signed Windows driver.

Signed-off-by: Lorenz Bauer <lmb@isovalent.com>