link: don't crash when calling testLink with a RawLink directly
testLink still has the assumption that calling RawLink.Info will
return an Info appropriate to the underlying link type. This is
not the case anymore since a recent refactoring. Instead switch
on the type of the link.

Another problem is that the test assumes that pinning and loading
from a pin as well as creating a link from an FD will return the
same type. This isn't the case when pinning a RawLink. Relax the
assertion.

Signed-off-by: Lorenz Bauer <lmb@isovalent.com>