link: minimal windows support WIP
We retain the Link interface and the RawLink type, but all other link
types are not available on Windows. There is a possibility that we can
extend the XDP link to Windows, but that will have to come later.

There is no support to retrieve link information or to pass additional
context to the link create call.

- Can probably add a translation layer for BPF_LINK_CREATE

Signed-off-by: Lorenz Bauer <lmb@isovalent.com>