internal: add a way to write portable feature tests
Feature tests are used to enable new features in a backwards compatible
manner. They exercise very specific code paths in the kernel. It makes
sense to assume that the existing feature tests will not work as intended
when porting to a new platform.

Change the NewFeatureTest constructor to take a list of versions, which
can be used to specify which operating systems are supported.

Signed-off-by: Lorenz Bauer <lmb@isovalent.com>