CI: switch arm64 runners to GHA
Seems like the CNCF is stopping funding for the external arm64 runners.
Switch to the GitHub provided ones.

For some reason TestGoarches/loong64 fails on these runners:

        error:
          got non-nil value
        comment:
          go output is:
          go: unsupported GOOS/GOARCH pair linux/loong64
        got:
          e"exit status 2"
        stack:
          /home/runner/work/ebpf/ebpf/cmd/bpf2go/gen/target_test.go:117
            qt.Assert(t, qt.IsNil(err), qt.Commentf("go output is:\n%s", string(output)))

The failing command is

    GOROOT=/ GOOS=linux GOARCH=loong64 go env

but I've been unable to reproduce the issue locally and in CI when running
outside of the test suite. Skip the test on arm64 for now.

Signed-off-by: Lorenz Bauer <lmb@isovalent.com>