prog, map: add nil checks in value marshalers
In case of programs or maps as map value types, the value marshaling
results in a panic in case of a nil pointer. For other types this is not a
problem since sysenc.Marshal does a nil check.

This commit adds a nil check to the internal helper for marshaling
programs and maps.

Signed-off-by: Tobias Böhm <code@aibor.de>