regenerate all TARGETS in main
This commit simply runs `make` to regenerate all the bpf2go targets in the
cmd, docs, and examples directories. This is useful so that in the upcoming
commit we can add a simple test in `cmd/bpf2go/test/test_bpfe{b,l}.go` to
make sure that a Variable/VariableSpec is emitted and loaded.

Signed-off-by: Simone Magnani <simone.magnani@isovalent.com>