add bpf2go test emit-and-load specs+bpf objects
This commit extends `api_test.go` to check whether specs and bpf objects
are correctly loaded. It also accounts for the new Variable/VariableSpec
in the test.
In addition, this commit adds a bpf2go test in `output_test.go` to verify
that all the listed bpf objects (i.e., variables, maps and programs) are
emitted. The tests looks for the output program, stored in a buffer, which
should contain the expected declarations.

This commits introduces further checks in the `api_test.go`

Signed-off-by: Simone Magnani <simone.magnani@isovalent.com>