variable: return *btf.Var from Variable{Spec}.Type()
In the initial design, we decided to return the inner type from .Type(), but
I forgot the outer btf.Var was needed to access decl tags set on the variable.

Returning a *Var here is the most convenient way of accessing the tags, otherwise
the caller would need to find the underlying map and parse the datasec manually,
which goes against what this API aims to solve.

Instead of the initial idea of returning a btf.Var as btf.Type, this now returns
btf.Var directly.

Signed-off-by: Timo Beckers <timo@isovalent.com>