variable: add VariableSpec.MapName() and .Offset() methods
Cilium's variable inliner needs this information to look up variable contents
based on the offset specified in the instruction stream, and wants to validate
that it's pulling data from the right data section.

It makes for a slightly leaky abstraction, but at least something meaningful
can be accomplished by combining the two, e.g. by accessing the given offset
in the map.

Signed-off-by: Timo Beckers <timo@isovalent.com>