elf: include data sections without references in CollectionSpec
Before VariableSpec, this was a nice optimization since no symbol references
means no ProgramSpec could ever refer to the map, effectively making it dead
code. There was a clear benefit to skipping it, although in practice having
a datasec without references should be quite rare.

With VariableSpec in the picture, the contents of a data section now affect
things like code generation through bpf2go. If an object contains variable
declarations like global `volatile const` and all of its references are
compiled out using compile-time flags, it won't show up in CollectionSpec.Variables
if its data section was skipped due to having no references.

While migrating the Cilium agent to using VariableSpec, I noticed some
permutations of build flags caused all references to a particular datasec to
drop, omitting it from spec.Variables. This patch removes the datasec pruning,
making the contents of spec.Variables more deterministic in the face of changing
build-time environments.

For users of NewCollection(), this can create and populate a .rodata map
unnecessarily, but if it's not referred to by any programs and the Collection
is closed, the map will be freed.

For users of LoadAndAssign(), the loading behaviour doesn't change since the
map won't be pulled in for lazy-loading if there aren't any program references.

Signed-off-by: Timo Beckers <timo@isovalent.com>