btf: replace decl and type tags with placeholders on unsupported kernels
Decl tags and type tags appeared in 5.16 and 5.17 respectively. This patch
makes it possible to load BTF containing such tags on older kernels by
replacing decl tags with integers and type tags with const qualifiers. This
mirrors the logic in libbpf.

This is convenient if the user wants to load an ELF containing decl tags on
global variables (available through CollectionSpec.Variables[].Tags) on older
kernels.

Signed-off-by: Timo Beckers <timo@isovalent.com>