info: add tests for recent ProgramInfo API additions
Added tests for:
- FuncInfos()
- LineInfos()
- JitedKsymAddrs()
- JitedInsns()
- JitedLineInfos()
- JitedFuncLens()

The existing test for JitedKsymAddrs was removed because it's not strictly
necessary to load a CollectionSpec to exercise it. It obscured the fact that
it seems like kernels before at least 5.4 need to have a subprog for jited
addrs to be returned in prog info.

TestProgInfoFuncInfos was removed for a similar reason. It's possible to
craft a program using the asm package that meets the criteria for exercising
FuncInfos() and LineInfos() in a deterministic way, no need to load a Collection.

Signed-off-by: Timo Beckers <timo@isovalent.com>