catch sys.ENOTSUPP error in probeProgram
This commit adds the sys.ENOTSUPP error to our error catching logic in
probeProgram. In the function, we then need to differentiate whether we
receive the error with an ebpf.StructOps or not: in the former case,
it means the program type is at least known to the kernel, while the
latter can occur anytime that, for instance, the JIT fails. When JIT fails,
the error returned ENOTSUPP is different than what we expect so far (EINVAL).

Signed-off-by: Simone Magnani <simone.magnani@isovalent.com>