features: consider ENOTSUPP in v2 and v4 ISA probes as conclusive
With the addition of the v4 ISA probe, we noticed an ENOTSUPP being returned by
BPF_PROG_LOAD on aarch64. Typically, this is an error code internal to the
verifier/JIT, but on aarch64 it seems to bubble up.

Since we cannot currently test older arm kernels, apply this logic to the older
ISA probes as well.

Signed-off-by: Timo Beckers <timo@isovalent.com>