prog: add ProgramOptions.LogSizeStart to obtain full log after verifier bug
When the verifier hits a bug, it returns EFAULT while clobbering other retcodes.
This defeats the buffer growth algorithm, so the user needs a break-glass
procedure for obtaining the full verifier log.

Also, reinstate maxVerifierLogSize since the maxu32 >> 2 limit is still in effect
in the Linux kernel. This can be removed if the limit is removed one day and
someone has a verifier log that's actually this big.

EFAULT can't be reliably triggered in CI, and I'm currently out of ideas for
testing this.

Signed-off-by: Timo Beckers <timo@isovalent.com>