ebpf: encode platform in type constants
Linux and Windows share the concept of program types, map types and so on.
For example, XDP is supported on both platforms. Unfortunately the constant
values used by both platforms are different. On Linux, XDP has value 0x6
while on Windows it is 0x1.

This problem extends to a CollectionSpec parsed from an ELF. Because of the
different values a Windows PerCPUHash aliases with a Linux ProgramArray.
This causes all sorts of mayhem that is hard to avoid without making too
many changes to the code base.

Introduce a Platform type which enumerates the supported platforms.
That type is then used to tag constant values. This means that there are
distinct constants even when a type exists on both platforms. Instead of
overloading the XDP constant we will introduce WindowsXDP.
This will allow the ELF reader to remain fully deterministic as well:
we can parse Linux ELF on Windows and vice versa.

The value of the Platform tag is chosen so that on Linux the constant
values do not change and are identical with the value in upstream
headers. On Windows the constants have a fixed (but implementation
defined) offset.

The downside of this approach is that Windows constants will all require
a prefix of some sort to distinguish them from Linux ones. This is probably
most onerous for map types, because those tend to be created manually more
frequently than programs, for example. We can add some behind the scenes
translation of Linux map types to Windows ones if need be.

Another downside is that constant values are now limited to 28 bits since
we steal the top 4 bits to store the platform. The constants we're applying
this to are all sequentially numbered from 0, so this is hopefully enough.

Signed-off-by: Lorenz Bauer <lmb@isovalent.com>